// auth-service/src/permission/enum/permission.enum.ts
export enum PermissionEnum {
  MANAGE_USER = 'manage_user',
  CREATE_USER = 'create_user',
  READ_USER = 'read_user',
  UPDATE_USER = 'update_user',
  DELETE_USER = 'delete_user',

  MANAGE_ROLE = 'manage_role',
  CREATE_ROLE = 'create_role',
  READ_ROLE = 'read_role',
  UPDATE_ROLE = 'update_role',
  DELETE_ROLE = 'delete_role',

  MANAGE_FILE = 'manage_file',
  CREATE_FILE = 'create_file',
  READ_FILE = 'read_file',
  UPDATE_FILE = 'update_file',
  DELETE_FILE = 'delete_file',

  MANAGE_PERMISSION = 'manage_permission',
  CREATE_PERMISSION = 'create_permission',
  READ_PERMISSION = 'read_permission',
  UPDATE_PERMISSION = 'update_permission',
  DELETE_PERMISSION = 'delete_permission',

  MANAGE_PROFILE = 'manage_profile',
  READ_PROFILE = 'read_profile',
  UPDATE_PROFILE = 'update_profile',

  MANAGE_PROFILE_IMAGE = 'manage_profileImage',
  CREATE_PROFILE_IMAGE = 'create_profileImage',
  DELETE_PROFILE_IMAGE = 'delete_profileImage',
}